requirements.txt
streamlit>=1.28.0
grpcio>=1.50.0
protobuf>=4.21.0
requests>=2.28.0 